+++ src/api/films-api-service.js
@@ -1,10 +1,4 @@
-import ApiService from './framework/api-service';
-import { TransformKeysObject } from './util';
-
-const Method = {
-  GET: 'GET',
-  PUT: 'PUT',
-};
+import ApiService from '../framework/api-service';
 
 export default class FilmsApiService extends ApiService {
   get films() {
@@ -14,18 +8,12 @@ export default class FilmsApiService extends ApiService {
   updateFilm = async (film) => {
     const response = await this._load({
       url: `movies/${film.id}`,
-      method: Method.PUT,
-      body: JSON.stringify(this.#adaptToServer(film)),
+      method: this.Method.PUT,
+      body: JSON.stringify(ApiService.adaptToServer(film)),
       headers: new Headers({ 'Content-Type': 'application/json' }),
     });
     const parsedResponse = await ApiService.parseResponse(response);
 
     return parsedResponse;
-  };
-
-  #adaptToServer = (film) => {
-    const adaptedFilm = new TransformKeysObject(film).fromCamelToSnakeCase();
-
-    return adaptedFilm;
   };
 }
